Factor out SwiftDWARFImporterDelegate into a separate file (NFC)
