+++ lldb/source/Plugins/TypeSystem/Swift/SwiftASTContext.cpp
@@ -57,8 +57,6 @@
 #include "swift/Serialization/Validation.h"
 #include "swift/SymbolGraphGen/SymbolGraphOptions.h"
 #include "clang/AST/ASTContext.h"
-#include "clang/AST/DeclObjC.h"
-#include "clang/Basic/SourceManager.h"
 #include "clang/Basic/TargetInfo.h"
 #include "clang/Basic/TargetOptions.h"
 #include "clang/Driver/Driver.h"
@@ -89,7 +87,6 @@
 #include "swift/Strings.h"
 
 #include "Plugins/ExpressionParser/Clang/ClangHost.h"
-#include "Plugins/ExpressionParser/Clang/ClangUtil.h"
 #include "Plugins/ExpressionParser/Swift/SwiftDiagnostic.h"
 #include "Plugins/ExpressionParser/Swift/SwiftHost.h"
 #include "Plugins/ExpressionParser/Swift/SwiftUserExpression.h"
@@ -110,7 +107,6 @@
 #include "lldb/Symbol/ObjectFile.h"
 #include "lldb/Symbol/SourceModule.h"
 #include "lldb/Symbol/SymbolFile.h"
-#include "lldb/Symbol/TypeMap.h"
 #include "lldb/Symbol/VariableList.h"
 #include "lldb/Target/Platform.h"
 #include "lldb/Target/Process.h"
@@ -119,7 +115,6 @@
 #include "lldb/Utility/FileSpec.h"
 #include "lldb/Utility/LLDBAssert.h"
 #include "lldb/Utility/LLDBLog.h"
-#include "lldb/Utility/Log.h"
 #include "lldb/Utility/ReproducerProvider.h"
 #include "lldb/Utility/Status.h"
 #include "lldb/Utility/Timer.h"
@@ -1950,7 +1945,7 @@ static lldb::ModuleSP GetUnitTestModule(lldb_private::ModuleList &modules) {
   return ModuleSP();
 }
 
-static SwiftASTContext *GetModuleSwiftASTContext(Module &module) {
+SwiftASTContext *lldb_private::GetModuleSwiftASTContext(Module &module) {
   auto type_system_or_err =
       module.GetTypeSystemForLanguage(lldb::eLanguageTypeSwift);
   if (!type_system_or_err) {
@@ -3038,283 +3033,6 @@ private:
   bool m_colorize;
 };
 
-/// Implements a swift::DWARFImporterDelegate to look up Clang types in DWARF.
-///
-/// During compile time, ClangImporter-imported Clang modules are compiled with
-/// -gmodules, which emits a DWARF rendition of all types defined in the module
-/// into the .pcm file. On Darwin, these types can be collected by
-/// dsymutil. This delegate allows DWARFImporter to ask LLDB to look up a Clang
-/// type by name, synthesize a Clang AST from it. DWARFImporter then hands this
-/// Clang AST to ClangImporter to import the type into Swift.
-class SwiftDWARFImporterDelegate : public swift::DWARFImporterDelegate {
-  SwiftASTContext &m_swift_ast_ctx;
-  using ModuleAndName = std::pair<const char *, const char *>;
-  std::string m_description;
-
-  /// Used to filter out types with mismatching kinds.
-  bool HasTypeKind(TypeSP clang_type_sp, swift::ClangTypeKind kind) {
-    CompilerType fwd_type = clang_type_sp->GetForwardCompilerType();
-    clang::QualType qual_type = ClangUtil::GetQualType(fwd_type);
-    switch (kind) {
-    case swift::ClangTypeKind::Typedef:
-      /*=swift::ClangTypeKind::ObjCClass:*/
-      return !qual_type->isObjCObjectOrInterfaceType() &&
-             !qual_type->getAs<clang::TypedefType>();
-    case swift::ClangTypeKind::Tag:
-      return !qual_type->isStructureOrClassType() &&
-             !qual_type->isEnumeralType() && !qual_type->isUnionType();
-    case swift::ClangTypeKind::ObjCProtocol:
-      // Not implemented since Objective-C protocols aren't yet
-      // described in DWARF.
-      return true;
-    }
-  }
-
-  clang::Decl *GetDeclForTypeAndKind(clang::QualType qual_type,
-                                     swift::ClangTypeKind kind) {
-    switch (kind) {
-    case swift::ClangTypeKind::Typedef:
-      /*=swift::ClangTypeKind::ObjCClass:*/
-      if (auto *obj_type = qual_type->getAsObjCInterfaceType())
-        return obj_type->getInterface();
-      if (auto *typedef_type = qual_type->getAs<clang::TypedefType>())
-        return typedef_type->getDecl();
-      break;
-    case swift::ClangTypeKind::Tag:
-      return qual_type->getAsTagDecl();
-    case swift::ClangTypeKind::ObjCProtocol:
-      // Not implemented since Objective-C protocols aren't yet
-      // described in DWARF.
-      break;
-    }
-    return nullptr;
-  }
-
-  static CompilerContextKind
-  GetCompilerContextKind(llvm::Optional<swift::ClangTypeKind> kind) {
-    if (!kind)
-      return CompilerContextKind::AnyType;
-    switch (*kind) {
-    case swift::ClangTypeKind::Typedef:
-      /*=swift::ClangTypeKind::ObjCClass:*/
-      return (CompilerContextKind)((uint16_t)CompilerContextKind::Any |
-                                   (uint16_t)CompilerContextKind::Typedef |
-                                   (uint16_t)CompilerContextKind::Struct);
-      break;
-    case swift::ClangTypeKind::Tag:
-      return (CompilerContextKind)((uint16_t)CompilerContextKind::Any |
-                                   (uint16_t)CompilerContextKind::Class |
-                                   (uint16_t)CompilerContextKind::Struct |
-                                   (uint16_t)CompilerContextKind::Union |
-                                   (uint16_t)CompilerContextKind::Enum);
-      // case swift::ClangTypeKind::ObjCProtocol:
-      // Not implemented since Objective-C protocols aren't yet
-      // described in DWARF.
-    default:
-      return CompilerContextKind::Invalid;
-    }
-  }
-
-  /// Import \p qual_type from one clang ASTContext to another and
-  /// add it to \p results if successful.
-  void importType(clang::QualType qual_type, clang::ASTContext &from_ctx,
-                  clang::ASTContext &to_ctx,
-                  llvm::Optional<swift::ClangTypeKind> kind,
-                  llvm::SmallVectorImpl<clang::Decl *> &results) {
-    clang::ASTImporter importer(to_ctx,
-                                to_ctx.getSourceManager().getFileManager(),
-                                from_ctx,
-                                from_ctx.getSourceManager().getFileManager(),
-                                false);
-    llvm::Expected<clang::QualType> clang_type(importer.Import(qual_type));
-    if (!clang_type) {
-      llvm::consumeError(clang_type.takeError());
-      return;
-    }
-
-    // Retrieve the imported type's Decl.
-    if (kind) {
-      if (clang::Decl *clang_decl = GetDeclForTypeAndKind(*clang_type, *kind))
-        results.push_back(clang_decl);
-    } else {
-      swift::ClangTypeKind kinds[] = {
-          swift::ClangTypeKind::Typedef, // =swift::ClangTypeKind::ObjCClass,
-          swift::ClangTypeKind::Tag, swift::ClangTypeKind::ObjCProtocol};
-      for (auto kind : kinds)
-        if (clang::Decl *clang_decl = GetDeclForTypeAndKind(*clang_type, kind))
-          results.push_back(clang_decl);
-    }
-  }
-
-public:
-  SwiftDWARFImporterDelegate(SwiftASTContext &swift_ast_ctx)
-      : m_swift_ast_ctx(swift_ast_ctx),
-        m_description(swift_ast_ctx.GetDescription() +
-                      "::SwiftDWARFImporterDelegate") {}
-
-  /// Look up a clang::Decl by name.
-  ///
-  /// There are two primary ways that this delegate method is called:
-  ///
-  ///    1. When resolving a type from a mangled name. In this case \p
-  ///       kind will be known, but the owning module of a Clang type
-  ///       in a mangled name is always __ObjC or __C.
-  ///
-  ///    2. When resolving a type from a serialized module
-  ///       cross reference. In this case \c kind will be unspecified,
-  ///       but the (top-level) module that the type is defined in
-  ///       will be known.
-  ///
-  /// The following diagram shows how the various components
-  /// interact. All paths lead to a call to the function
-  /// \c ClangImporter::Implementation::importDeclReal(), which turns
-  /// a \c clang::Decl into a \c swift::Decl.  The return paths leading
-  /// back from \c importDeclReal() are omitted from the diagram. Also
-  /// some auxiliary intermediate function calls are be omitted for
-  /// brevity.
-  ///
-  /// \verbatim
-  /// ╔═LLDB═════════════════════════════════════════════════════════════════╗
-  /// ║                                                                      ║
-  /// ║  ┌─DWARFASTParserSwift──────────┐   ┌─DWARFImporterDelegate──────┐   ║
-  /// ║  │                              │   │                            │   ║
-  /// ║  │ GetTypeFromMangledTypename() │ ┌─├→lookupValue()─────┐        │   ║
-  /// ║  │             │                │ │ │                   │        │   ║
-  /// ║  └─────────────┬────────────────┘ │ └───────────────────┬────────┘   ║
-  /// ║                │                  │                     │            ║
-  /// ╚════════════════╤══════════════════╧═════════════════════╤════════════╝
-  ///                  │                  │                     │
-  /// ╔═Swift Compiler═╤══════════════════╧═════════════════════╤════════════╗
-  /// ║                │                  │                     │            ║
-  /// ║  ┌─ASTDemangler┬─────────────┐    │ ┌─ClangImporter─────┬────┐       ║
-  /// ║  │             ↓             │    │ │                   │    │       ║
-  /// ║  │ findForeignTypeDecl()─────├──────├→lookupTypeDecl()  │    │       ║
-  /// ║  │                           │    │ │      ⇣            ↓    │       ║
-  /// ║  └───────────────────────────┘    └─┤─lookupTypeDeclDWARF()  │       ║
-  /// ║                                     │      ↓                 │       ║
-  /// ║                                     │ *importDeclReal()*     │       ║
-  /// ║                                     │      ↑                 │       ║
-  /// ║                                     │ lookupValueDWARF()     │       ║
-  /// ║                                     │      ↑                 │       ║
-  /// ║                                     └──────┴─────────────────┘       ║
-  /// ║                                            │                         ║
-  /// ║  ┌─Deserialization─────────┐               └──────────────────────┐  ║
-  /// ║  │ loadAllMembers()        │                                      │  ║
-  /// ║  │        ↓                │  ┌─ModuleDecl────┐ ┌─DWARFModuleUnit─┴┐ ║
-  /// ║  │ resolveCrossReference()─├──├→lookupValue()─├─├→lookupValue()───┘│ ║
-  /// ║  │                         │  └───────────────┘ └──────────────────┘ ║
-  /// ║  └─────────────────────────┘                                         ║
-  /// ╚══════════════════════════════════════════════════════════════════════╝
-  /// \endverbatim
-  void lookupValue(StringRef name, llvm::Optional<swift::ClangTypeKind> kind,
-                   StringRef inModule,
-                   llvm::SmallVectorImpl<clang::Decl *> &results) override {
-    LLDB_SCOPED_TIMER();
-    LOG_PRINTF(GetLog(LLDBLog::Types), "(\"%s\")", name.str().c_str());
-
-    // We will not find any Swift types in the Clang compile units.
-    ConstString name_cs(name);
-    if (SwiftLanguageRuntime::IsSwiftMangledName(name_cs.GetStringRef()))
-      return;
-
-    auto clang_importer = m_swift_ast_ctx.GetClangImporter();
-    if (!clang_importer)
-      return;
-
-    // Find the type in the debug info.
-    TypeMap clang_types;
-    ConstString module_cs(inModule);
-
-    llvm::SmallVector<CompilerContext, 3> decl_context;
-    // Perform a lookup in a specific module, if requested.
-    if (!inModule.empty())
-      decl_context.push_back({CompilerContextKind::Module, module_cs});
-    // Swift doesn't keep track of submodules.
-    decl_context.push_back({CompilerContextKind::AnyModule, ConstString()});
-    decl_context.push_back({GetCompilerContextKind(kind), name_cs});
-    llvm::DenseSet<SymbolFile *> searched_symbol_files;
-    auto search = [&](Module &module) {
-      module.FindTypes(decl_context,
-                       TypeSystemClang::GetSupportedLanguagesForTypes(),
-                       searched_symbol_files, clang_types);
-    };
-    if (Module *module = m_swift_ast_ctx.GetTypeSystemSwiftTypeRef().GetModule())
-      search(*module);
-    else if (TargetSP target_sp = m_swift_ast_ctx.GetTargetWP().lock()) {
-      // In a scratch context, check the module's DWARFImporterDelegates first.
-      //
-      // It's a common pattern that a type is revisited immediately
-      // after looking it up in a per-module context in the scratch
-      // context for dynamic type resolution.
-      auto images = target_sp->GetImages();
-      for (size_t i = 0; i != images.GetSize(); ++i) {
-        auto module_sp = images.GetModuleAtIndex(i);
-        auto *swift_ast_ctx = GetModuleSwiftASTContext(*module_sp);
-	if (!swift_ast_ctx)
-          continue;
-        auto *dwarf_imp = static_cast<SwiftDWARFImporterDelegate *>(
-            swift_ast_ctx->GetDWARFImporterDelegate());
-        if (!dwarf_imp || dwarf_imp == this)
-          continue;
-
-        llvm::SmallVector<clang::Decl *, 2> module_results;
-        dwarf_imp->lookupValue(name, kind, inModule, module_results);
-        if (!module_results.size())
-          continue;
-
-        auto *from_clang_importer = swift_ast_ctx->GetClangImporter();
-        if (!from_clang_importer)
-          continue;
-        auto &from_ctx = from_clang_importer->getClangASTContext();
-        auto &to_ctx = clang_importer->getClangASTContext();
-        for (clang::Decl *decl : module_results) {
-          clang::QualType qual_type;
-          if (auto *interface = llvm::dyn_cast<clang::ObjCInterfaceDecl>(decl))
-            qual_type = {interface->getTypeForDecl(), 0};
-          if (auto *type = llvm::dyn_cast<clang::TypeDecl>(decl))
-            qual_type = {type->getTypeForDecl(), 0};
-          importType(qual_type, from_ctx, to_ctx, kind, results);
-        }
-        // Cut the search short after we found the first result.
-        if (results.size())
-          break;
-      }
-      LOG_PRINTF(GetLog(LLDBLog::Types), "%zu types collected.",
-                 results.size());
-      return;
-    }
-
-    clang_types.ForEach([&](lldb::TypeSP &clang_type_sp) {
-      if (!clang_type_sp)
-        return true;
-
-      // Filter out types with a mismatching type kind.
-      if (kind && HasTypeKind(clang_type_sp, *kind))
-        return true;
-
-      // Realize the full type.
-      CompilerType compiler_type = clang_type_sp->GetFullCompilerType();
-
-      // Filter our non-Clang types.
-      auto *type_system = llvm::dyn_cast_or_null<TypeSystemClang>(
-          compiler_type.GetTypeSystem());
-      if (!type_system)
-        return true;
-
-      // Import the type into the DWARFImporter's context.
-      clang::ASTContext &to_ctx = clang_importer->getClangASTContext();
-      clang::ASTContext &from_ctx = type_system->getASTContext();
-
-      clang::QualType qual_type = ClangUtil::GetQualType(compiler_type);
-      importType(qual_type, from_ctx, to_ctx, kind, results);
-
-      return true;
-    });
-
-    LOG_PRINTF(GetLog(LLDBLog::Types), "%zu types from debug info.",
-               results.size());
-  }
-};
 } // namespace lldb_private
 
 swift::ASTContext *SwiftASTContext::GetASTContext() {
@@ -3349,8 +3067,8 @@ swift::ASTContext *SwiftASTContext::GetASTContext() {
       // Create the DWARFImporterDelegate.
       const auto &props = ModuleList::GetGlobalModuleListProperties();
       if (props.GetUseSwiftDWARFImporter())
-        m_dwarf_importer_delegate_up =
-            std::make_unique<SwiftDWARFImporterDelegate>(*this);
+        m_dwarf_importer_delegate_up.reset(
+            CreateSwiftDWARFImporterDelegate(*this));
       clang_importer_ap = swift::ClangImporter::create(
           *m_ast_context_ap, "",
           m_dependency_tracker.get(), m_dwarf_importer_delegate_up.get());
